Use base model from configuration as basis of future metadata containers
Eventually, we should have as little custom code as possible, leaving
`DataContainer` and custom machinery of validation and normalization
behind us, switching to attrs and Pydantic. What started in
`tmt.config.models`, we can use in general for containers populated by
fmf and/or CLI.
