--- tmt/config/models/__init__.py
@@ -1,10 +1,0 @@
-from tmt._compat.pydantic import BaseModel, Extra
-from tmt.container import key_to_option
-
-
-class BaseConfig(BaseModel):
-    class Config:
-        # Accept only keys with dashes instead of underscores
-        alias_generator = key_to_option
-        extra = Extra.forbid
-        validate_assignment = True
