+++ tmt/config/models/link.py
@@ -1,20 +1,19 @@
 from enum import Enum
 
 from tmt._compat.pydantic import HttpUrl
-
-from . import BaseConfig
+from tmt.container import MetadataContainer
 
 
 class IssueTrackerType(str, Enum):
     jira = 'jira'
 
 
-class IssueTracker(BaseConfig):
+class IssueTracker(MetadataContainer):
     type: IssueTrackerType
     url: HttpUrl
     tmt_web_url: HttpUrl
     token: str
 
 
-class LinkConfig(BaseConfig):
+class LinkConfig(MetadataContainer):
     issue_tracker: list[IssueTracker]
